company_scraper/pyproject.toml
[tool.poetry]
name = "company-scraper"
version = "0.1.0"
description = ""
authors = ["Joshua Westphal da Silva <joshuasilva333@gmail.com>"]
readme = "README.md"
packages = [{include = "company_scraper"}]

[tool.poetry.dependencies]
python = "^3.10"
black = "^23.1.0"
fire = "^0.5.0"
sqlitedict = "^2.1.0"
Scrapy = "^2.8.0"
python-dateutil = "^2.8.2"
scrapy-rotating-proxies = "^0.6.2"
scrapy-fake-useragent = "^1.4.4"
docker-compose = "^1.29.2"
mongo = "^0.2.0"
mongoengine = "^0.27.0"
slugify = "^0.0.1"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
